# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo Web para la Centralización de Documentos Electrónicos Tributarios",
        "contact": {
            "name": "Rafael Larrea",
            "email": "jrafael1108@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rol/list/{estado}": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "Obtiene una lista de roles por estado.",
                "operationId": "14ea5429b3d23cdb9c4cda4e0da2ac58",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "path",
                        "description": "Filtra el rol por estado.",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
...
        "/api/rol/{id}": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "Obtener detalles de un rol por ID",
                "operationId": "da95f4131c7a9793e4a5692f5f53801e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la rol"
                    },
                    "404": {
                        "description": "rol no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rol"
                ],
                "summary": "Actualiza la entidad rol",
                "operationId": "912366a682e11a5503fbd4df27e5a79d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de rol a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Nombre": {
                                        "type": "string"
                                    },
                                    "Descripcion": {
                                        "type": "string"
                                    },
                                    "Permisos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "IdUsuario": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/rol/permisos/": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "Listado de permisos",
                "operationId": "a4ebadffa9a6310e1a5488fac8a43045",
                "responses": {
                    "200": {
                        "description": "Listado detalle de permiso"
                    },
                    "404": {
                        "description": "Permiso no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/rol/": {
            "post": {
                "tags": [
                    "Rol"
                ],
                "summary": "Crear una nuevo rol",
                "operationId": "81802298c4930ac383cc457d559ea7a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Nombre": {
                                        "type": "string"
                                    },
                                    "Descripcion": {
                                        "type": "string"
                                    },
                                    "Permisos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "IdUsuario": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/rol/deshabilitar/{id}": {
            "post": {
                "tags": [
                    "Rol"
                ],
                "summary": "Deshabilita una entidad rol",
                "operationId": "70f124142b88d57d0f242c7ca5619bf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la rol a deshabilitar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/rol/habilitar/{id}": {
            "post": {
                "tags": [
                    "Rol"
                ],
                "summary": "Habilita una entidad rol",
                "operationId": "65de27a3b0ec5d0590dd24f15062dca2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la rol a habilitar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIResponse": {
                "description": "",
                "properties": {
                    "Success": {
                        "type": "boolean"
                    },
                    "Message": {
                        "type": "string"
                    },
                    "Data": {
                        "type": "object"
                    },
                    "Exception": {
                        "type": "object"
                    },
                    "Code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "in": "header"
            },
            "Session": {
                "type": "apiKey",
                "description": "Enter your session id",
                "name": "SessionId",
                "in": "header"
            },
            "Apikey": {
                "type": "apiKey",
                "description": "Enter your APIKEY for company",
                "name": "APIKEY",
                "in": "header"
            }
        }
    },
...
}